server/requirements.txt
Flask
tensorflow
numpy
opencv-python-headless
imutils
scikit-learn
joblib
Pillow
flask-cors